+++ src/converter.js
@@ -9,6 +9,8 @@ const JSON_SCHEMA_TO_BIGQUERY_TYPE_DICT = {
   string: 'STRING'
 }
 
+const OFS = ['allOf', 'anyOf', 'oneOf']
+
 converter._merge_property = (merge_type, property_name, destination_value, source_value) => {
   //Merges two properties.
   let destination_list
@@ -21,9 +23,9 @@ converter._merge_property = (merge_type, property_name, destination_value, sourc
   if(source_value === undefined){
     return destination_value
   }
-	if(typeof destination_value === 'boolean' && typeof source_value === 'boolean'){
-		return destination_value && source_value
-	}
+  if(typeof destination_value === 'boolean' && typeof source_value === 'boolean'){
+    return destination_value && source_value
+  }
   if(_.isPlainObject(destination_value) && _.isPlainObject(source_value)){
     return converter._merge_dicts(merge_type, destination_value, source_value)
   }
@@ -32,6 +34,7 @@ converter._merge_property = (merge_type, property_name, destination_value, sourc
   }else{
     destination_list = [destination_value]
   }
+  let source_list
   if(Array.isArray(source_value)){
     source_list = source_value
   }else{
@@ -56,20 +59,27 @@ converter._deepCopy = (o) => _.clone(o, true)
  * Cloned from the original merge_dicts from python-based bigjson
  * This method had variable expansion, varags  and two slightly different
  * usages which appeared to be incompatible.
- * 
- * Could be returned to a single method 
- * 
+ *
+ * Could be returned to a single method
+ *
  * See merge_dicts for the alternate call pattern
  */
 converter._merge_dicts_array = (merge_type, dest_dict, source_dicts) => { //Array was *dicts, merges multiple
   const result = converter._deepCopy(dest_dict)
-  for(let i = 0; i < source_dicts.length; i++){
-    const source_dict = source_dicts[i]
+  for(let source_dict of source_dicts){
+
+    // First check if we need to recurse and merge deeper results first
+    for(const x_of of OFS){
+      if(_.has(source_dict, x_of)){
+        source_dict = converter._merge_dicts_array(x_of, source_dict, source_dict[x_of])
+        delete source_dict[x_of]
+      }
+    }
+
     const keys = Object.keys(source_dict)
-    for(let j=0; j<keys.length;j++){
-      const name = keys[j]
+    for(const name of keys){
       const merged_property = converter._merge_property(merge_type, name, result[name], source_dict[name])
-			if(merged_property !== undefined){
+      if(merged_property !== undefined){
         result[name] = merged_property
       }
     }
@@ -81,18 +91,17 @@ converter._merge_dicts_array = (merge_type, dest_dict, source_dicts) => { //Arra
  * The original merge_dicts from python-based bigjson
  * This method had variable expansion, varags  and two slightly different
  * usages which appeared to be incompatible.
- * 
- * Could be returned to a single method 
- * 
+ *
+ * Could be returned to a single method
+ *
  * See merge_dicts_array above for the alternate call pattern
  **/
 converter._merge_dicts = (merge_type, dest_dict, source_dict) => { //Merges a single object
   const result = converter._deepCopy(dest_dict)
   const keys = Object.keys(source_dict)
-  for(let j=0; j<keys.length;j++){
-    const name = keys[j]
+  for(const name of keys){
     const merged_property = converter._merge_property(merge_type, name, result[name], source_dict[name])
-		if(merged_property !== undefined){
+    if(merged_property !== undefined){
       result[name] = merged_property
     }
   }
@@ -101,6 +110,9 @@ converter._merge_dicts = (merge_type, dest_dict, source_dict) => { //Merges a si
 
 converter._scalar = (name, type_, mode, description) => {
   const bigquery_type = JSON_SCHEMA_TO_BIGQUERY_TYPE_DICT[type_]
+  if(!bigquery_type){
+    throw new Error(`Invalid type given: ${type_} for '${name}'`)
+  }
 
   const result = {
     name: name,
@@ -115,7 +127,7 @@ converter._scalar = (name, type_, mode, description) => {
   return result
 }
 
-converter._array = (name, node, mode) => {
+converter._array = (name, node) => {
   let items_with_description = converter._deepCopy(node['items'])
   if(_.has(items_with_description,'description')){
     items_with_description['description'] = node['description']
@@ -124,9 +136,9 @@ converter._array = (name, node, mode) => {
 }
 
 converter._object = (name, node, mode) => {
-	if (node.additionalProperties !== false) {
-		throw new Error(`'object' type properties must have an '"additionalProperties": false' property:\n${JSON.stringify(node, null, 2)}`)
-	}
+  if (node.additionalProperties !== false) {
+    throw new Error(`'object' type properties must have an '"additionalProperties": false' property:\n${JSON.stringify(node, null, 2)}`)
+  }
   const required_properties = node['required'] || []
   const properties = node['properties']
 
@@ -150,7 +162,7 @@ converter._object = (name, node, mode) => {
 
 converter._simple = (name, type_, node, mode) => {
   if(type_ === 'array'){
-    return converter._array(name, node, mode)
+    return converter._array(name, node)
   }
   if(type_ === 'object'){
     return converter._object(name, node, mode)
@@ -167,11 +179,9 @@ converter._simple = (name, type_, node, mode) => {
 
 converter._visit = (name, node, mode='NULLABLE') => {
   let merged_node = node
-  const ofs = ['allOf', 'anyOf', 'oneOf']
-  for(x=0 ; x<ofs.length; x++){
-    const x_of = ofs[x]
-    if(_.has(node,x_of)){
-      merged_node = converter._merge_dicts_array(x_of, node, _.get(node,x_of))
+  for(const x_of of OFS){
+    if(_.has(node, x_of)){
+      merged_node = converter._merge_dicts_array(x_of, node, _.get(node, x_of))
       delete merged_node[x_of]
     }
   }
@@ -187,7 +197,7 @@ converter._visit = (name, node, mode='NULLABLE') => {
     }
     type_ = non_null_types[0]
   }
-  return converter._simple(name, type_, merged_node,  actual_mode)
+  return converter._simple(name, type_, merged_node, actual_mode)
 }
 
 converter.run = (input_schema) => {